Don't re-run argon config

+++ bin/setup/case-argon.sh
@@ -12,9 +12,11 @@ download_argon() {
 }
 
 install() {
-  download_argon
-  bash "$argon_bin"
-  rm "$argon_bin"
+  if [ ! `command -v argonone-config` ]; then
+    download_argon
+    bash "$argon_bin"
+    rm "$argon_bin"
+  fi
 }
 
 uninstall() {
